pyproject.toml
[tool.poetry]
name = "perdiem-django"
version = "0.1.0"
description = "Crowdfunding platform for emerging musicians."
authors = ["Lucas Connors <lucas@revolutiontech.ca>", "localastronaut"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.6"
Django = "~2.2.4"
gunicorn = "^19.9.0"
psycopg2 = "=2.8.3"
Pillow = "^6.1.0"
django-classbasedsettings = "^1.5.0"
social-auth-app-django = "^3.1.0"
python-memcached = "^1.59"
requests = "^2.22.0"
boto3 = "^1.9.215"
django-pigeon = "^0.3.0"
sorl-thumbnail = "^12.5.0"
whitenoise = "^4.1.3"
django-s3-storage = "^0.12.5"
Fabric3 = "^1.14.post1"
djangorestframework = "^3.10.2"
raven = "^6.10.0"
django-ses = "^0.8.12"
geopy = "^1.20.0"
pinax-stripe = "^4.4.0"
django-templated-email = "^2.3.0"
django-markdown-deux = "^1.0.5"
django-pagedown = "^1.0.6"
django-gfklookupwidget = "^1.0.7"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
factory-boy = "^2.12.0"
coverage = "^4.5.4"
black = {version = "^18.3-alpha.0",allows-prereleases = true}

[tool.black]
line-length = 88
target_version = ["py36"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
